Move Form component to separate file
